requirements.txt
streamlit
openai
python-dotenv
Pillow
google-generativeai #for Gemini API integration